refactor: move /constants to /common folder
