+++ src/common/constants/cookies.ts
@@ -1,0 +1,7 @@
+import { DAY } from './time';
+
+export const ACCESS_TOKEN_KEY = '__schedly__accessToken';
+export const REFRESH_TOKEN_KEY = '__schedly__refreshToken';
+
+export const ACCESS_TOKEN_EXPIRES_IN = 7 * DAY;
+export const REFRESH_TOKEN_EXPIRES_IN = 30 * DAY;
