+++ src/common/constants/time.ts
@@ -1,0 +1,5 @@
+export const MILLISECOND = 1;
+export const SECOND = 1_000 * MILLISECOND;
+export const MINUTE = 60 * SECOND;
+export const HOUR = 60 * MINUTE;
+export const DAY = 24 * HOUR;
