--- src/utils/constants/cookies.ts
@@ -1,4 +1,0 @@
-export const ACCESS_TOKEN_KEY = '__schedly__accessToken';
-export const REFRESH_TOKEN_KEY = '__schedly__refreshToken';
-export const ACCESS_TOKEN_EXPIRES_IN = 7 * 24 * 60 * 60 * 1_000; // 7 days
-export const REFRESH_TOKEN_EXPIRES_IN = 30 * 24 * 60 * 60 * 1_000; // 30 days
